+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.7]
+
+### Fixed
+- Payment failed when the order line description exceeded 50 characters.
+- CapturedAmount is used instead of ReservedAmount in order payment setAmount.
+- Create a custom job to process notification requests asynchronously.
+- Set the order status to **Paid** only if the order has been captured.
+
 ## [2.0.6]
 
 ### Added